+++ packages/plugin-collapsible-list/web/src/tiptap/component.tsx
@@ -42,6 +42,7 @@ export const CollapsibleList: React.FC<PluginProps> = ({
       // onFocus={this.onFocus}
       // tabIndex="0"
       data-hook="collapsibleListComponent"
+      contentEditable="false"
     >
       <NodeViewContent />
       {inCollapsibleList && (
@@ -71,7 +72,13 @@ export const CollapsibleListItem: React.FC<PluginProps> = ({
   return (
     <div className={styles.itemContainer}>
       {inCollapsibleList && (
-        <div className={styles.dndIcon}>
+        <div
+          className={styles.dndIcon}
+          draggable="true"
+          contentEditable="false"
+          data-drag-handle
+          data-type="draggable-item"
+        >
           <DndIcon />
         </div>
       )}
@@ -95,7 +102,7 @@ export const CollapsibleListItem: React.FC<PluginProps> = ({
 export const CollapsibleListItemTitle: React.FC<PluginProps> = ({ NodeViewContent }) => {
   return (
     <div className={collapsibleListItemStyles.titleContainer}>
-      <NodeViewContent contenteditable="true" />
+      <NodeViewContent className={collapsibleListItemStyles.nodes} contenteditable="true" />
     </div>
   );
 };
@@ -114,7 +121,7 @@ export const CollapsibleListItemBody: React.FC<PluginProps> = ({
   )?.node?.attrs?.isExpanded;
   return isShown ? (
     <div className={collapsibleListItemStyles.titleContainer}>
-      <NodeViewContent contenteditable="true" />
+      <NodeViewContent className={collapsibleListItemStyles.nodes} contenteditable="true" />
     </div>
   ) : null;
 };